Create a mount point for my router's FTP server
The ASUS router has two USB ports that I occasionally use to
transfer files from the storage array to a USB thumb drive.

The router is configured to start an FTP server when a storage
device is plugged into one of its USB ports, and therefore the
server's availability is intermittent.

This commit creates a means of mounting the FTP service as a
local file system for the duration the thumb drive is plugged
into the router:

  # sudo mount /mnt/router
  # (transfer files to thumb drive)
  # sudo umount /mnt/router